implement ssh secrets
This change adds the capability of generating ssh secrets to the secret
operator.

While implementing this, I also moved the password generation to a
different file, which allows for further extension of the operator with
additional secret generators.

I also moved the test cases to generator specific files.

Furthermore, I noticed, that golangs string formatter on the time type
is quite verbose. The generated-at annotation now contains a
RFC3339 compliant time string.